requirements.txt
argparse
torch
pandas
plotly-express
scikit-learn
numpy
matplotlib